fix: read_heredoc.cの関数を分割
