+++ lexer/lexer.h
@@ -76,8 +76,12 @@ char			*expand_single_envvar(char *str, const char *from,
 					const char *to, size_t *i);
 char			*expand_envvar_str(const char *input, void *env);
 void			expand_envvar(t_list *lst, t_list *env_lst);
+void			join_line_to_buffer(char **buffer, char **line);
+
 char			*get_tmpfile_name(char *suffix);
 void			write_tmpfile(char *tmpfile, char *buf);
 void			create_tmpfile(char *tmpfile);
+void			write_heredocs(t_lexer *lexer, char *buf, char *delimiter);
+void			register_heredocs(t_lexer *lexer, char *delimiter);
 
 #endif
